requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
apify-client==2.4.1
schedule==1.2.0
